requirements.txt
python-dotenv
google-genai
fastapi
uvicorn
agno
pydantic
streamlit
pydantic[email]